+++ src/main/java/com/imsweb/datagenerator/naaccr/rule/patient/SexRule.java
@@ -5,6 +5,7 @@ import java.util.Map;
 import com.imsweb.datagenerator.naaccr.NaaccrDataGeneratorOptions;
 import com.imsweb.datagenerator.naaccr.NaaccrDataGeneratorPatientRule;
 import com.imsweb.datagenerator.utils.DistributionUtils;
+import com.imsweb.naaccrxml.NaaccrFormat;
 import com.imsweb.naaccrxml.entity.Patient;
 
 import static com.imsweb.datagenerator.naaccr.NaaccrDataGenerator.CONTEXT_FLAG_SEX;
@@ -23,6 +24,7 @@ public class SexRule extends NaaccrDataGeneratorPatientRule {
 
     @Override
     public void execute(Patient patient, NaaccrDataGeneratorOptions options, Map<String, Object> context) {
-        setValue(patient, "sex", (String)context.getOrDefault(CONTEXT_FLAG_SEX, DistributionUtils.getSex()));
+        String property = isOnOrBeforeNaaccrVersion(context, NaaccrFormat.NAACCR_VERSION_250) ? "sex" : "sexAssignedAtBirth";
+        setValue(patient, property, (String)context.getOrDefault(CONTEXT_FLAG_SEX, DistributionUtils.getSex()));
     }
 }