+++ src/test/java/lab/StagingDataGenerator.java
@@ -7,6 +7,7 @@ import java.io.BufferedReader;
 import java.io.BufferedWriter;
 import java.io.File;
 import java.io.IOException;
+import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.OutputStreamWriter;
 import java.nio.charset.StandardCharsets;
@@ -19,31 +20,24 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.regex.Pattern;
-import java.util.stream.Collectors;
 
 import org.apache.commons.lang3.StringUtils;
 
 import testing.TestingUtils;
 
 import com.imsweb.staging.Staging;
-import com.imsweb.staging.cs.CsDataProvider;
-import com.imsweb.staging.cs.CsDataProvider.CsVersion;
 import com.imsweb.staging.entities.ColumnDefinition.ColumnType;
 import com.imsweb.staging.entities.Input;
 import com.imsweb.staging.entities.Schema;
 import com.imsweb.staging.entities.SchemaLookup;
 import com.imsweb.staging.entities.Table;
-import com.imsweb.staging.eod.EodDataProvider;
-import com.imsweb.staging.eod.EodDataProvider.EodVersion;
-import com.imsweb.staging.tnm.TnmDataProvider;
-import com.imsweb.staging.tnm.TnmDataProvider.TnmVersion;
 
 public class StagingDataGenerator {
 
     public static void main(String[] args) throws Exception {
-        Staging cs = Staging.getInstance(CsDataProvider.getInstance(CsVersion.LATEST));
-        Staging tnm = Staging.getInstance(TnmDataProvider.getInstance(TnmVersion.LATEST));
-        Staging eod = Staging.getInstance(EodDataProvider.getInstance(EodVersion.LATEST));
+        Staging cs = loadStagingInstance("cs-02.05.50.zip");
+        Staging tnm = loadStagingInstance("tnm-2.1.zip");
+        Staging eod = loadStagingInstance("eod_public-3.3.zip");
 
         SchemaInfo info = new SchemaInfo();
 
@@ -58,8 +52,18 @@ public class StagingDataGenerator {
 
         File file = new File(TestingUtils.getWorkingDirectory() + "/src/main/resources/frequencies/staging_keys.csv");
         try (BufferedWriter writer = new BufferedWriter(new OutputStreamWriter(Files.newOutputStream(file.toPath()), StandardCharsets.UTF_8))) {
-            for (String key : keyCache.keySet().stream().sorted().collect(Collectors.toList()))
+            for (String key : keyCache.keySet().stream().sorted().toList()) {
                 writer.write(key + "," + keyCache.get(key) + "\r\n");
+            }
+        }
+    }
+
+    private static Staging loadStagingInstance(String data) {
+        try (InputStream is = Thread.currentThread().getContextClassLoader().getResourceAsStream("staging/" + data)) {
+            return Staging.getInstance(is);
+        }
+        catch (IOException e) {
+            throw new IllegalStateException("Unable to initialize staging from " + data, e);
         }
     }
 
@@ -113,29 +117,17 @@ public class StagingDataGenerator {
             newLines.add(hist);
             newLines.add(beh);
 
-            if (csSchema != null) {
-                info.getCsSchemas().add(csSchema.getId());
-                String key = keyCache.computeIfAbsent(csSchema.getId(), k -> String.valueOf(keyCache.size() + 1));
-                newLines.add(key);
-            }
-            else
-                newLines.add("");
+            info.getCsSchemas().add(csSchema.getId());
+            String key = keyCache.computeIfAbsent(csSchema.getId(), k -> String.valueOf(keyCache.size() + 1));
+            newLines.add(key);
 
-            if (tnmSchema != null) {
-                info.getTnmSchemas().add(tnmSchema.getId());
-                String key = keyCache.computeIfAbsent(tnmSchema.getId(), k -> String.valueOf(keyCache.size() + 1));
-                newLines.add(key);
-            }
-            else
-                newLines.add("");
+            info.getTnmSchemas().add(tnmSchema.getId());
+            key = keyCache.computeIfAbsent(tnmSchema.getId(), k -> String.valueOf(keyCache.size() + 1));
+            newLines.add(key);
 
-            if (eodSchema != null) {
-                info.getEodSchemas().add(eodSchema.getId());
-                String key = keyCache.computeIfAbsent(eodSchema.getId(), k -> String.valueOf(keyCache.size() + 1));
-                newLines.add(key);
-            }
-            else
-                newLines.add("");
+            info.getEodSchemas().add(eodSchema.getId());
+            key = keyCache.computeIfAbsent(eodSchema.getId(), k -> String.valueOf(keyCache.size() + 1));
+            newLines.add(key);
 
             lines.add(newLines);
         }
@@ -154,9 +146,11 @@ public class StagingDataGenerator {
         // only include a schema if it has at least one non-obsolete input (some TNM schema are completely obsolete)
         return schemas.stream()
                 .filter(schema -> {
-                    for (Input input : schema.getInputs().stream().sorted(Comparator.comparing(Input::getNaaccrXmlId)).collect(Collectors.toList()))
-                        if (!exclusions.contains(input.getKey()) && !input.getName().contains("OBSOLETE"))
+                    for (Input input : schema.getInputs().stream().sorted(Comparator.comparing(Input::getNaaccrXmlId)).toList()) {
+                        if (!exclusions.contains(input.getKey()) && !input.getName().contains("OBSOLETE")) {
                             return true;
+                        }
+                    }
                     return false;
                 })
                 .findFirst()
@@ -186,24 +180,27 @@ public class StagingDataGenerator {
         File file = new File(TestingUtils.getWorkingDirectory() + "/src/main/resources/" + filename);
         try (BufferedWriter writer = new BufferedWriter(new OutputStreamWriter(Files.newOutputStream(file.toPath()), StandardCharsets.UTF_8))) {
 
-            for (String schemaId : schemas.stream().sorted().collect(Collectors.toList())) {
+            for (String schemaId : schemas.stream().sorted().toList()) {
                 Schema schema = stating.getSchema(schemaId);
-                if (schema == null)
+                if (schema == null) {
                     throw new IllegalStateException("Unable to get schema '" + schemaId + "'");
+                }
 
                 Pattern digitRegx = Pattern.compile("^\\d+$|^\\d+-\\d+$");
 
-                for (Input input : schema.getInputs().stream().sorted(Comparator.comparing(Input::getNaaccrXmlId)).collect(Collectors.toList())) {
+                for (Input input : schema.getInputs().stream().sorted(Comparator.comparing(Input::getNaaccrXmlId)).toList()) {
                     if (!exclusions.contains(input.getKey()) && !input.getName().contains("OBSOLETE")) {
                         //System.out.println("\n***** " + input.getName());
                         Table table = stating.getTable(input.getTable());
 
                         int codeIdx = -1, descIdx = -1;
                         for (int i = 0; i < table.getColumnDefinitions().size(); i++) {
-                            if (table.getColumnDefinitions().get(i).getType() == ColumnType.INPUT)
+                            if (table.getColumnDefinitions().get(i).getType() == ColumnType.INPUT) {
                                 codeIdx = i;
-                            else if (table.getColumnDefinitions().get(i).getType() == ColumnType.DESCRIPTION)
+                            }
+                            else if (table.getColumnDefinitions().get(i).getType() == ColumnType.DESCRIPTION) {
                                 descIdx = i;
+                            }
                         }
 
                         if (codeIdx != -1) {
@@ -219,8 +216,9 @@ public class StagingDataGenerator {
                                     }
                                 }
                             }
-                            if (allDigitValues)
+                            if (allDigitValues) {
                                 codes = recombineRanges(codes);
+                            }
                             //System.out.println(input.getNaaccrXmlId() + ": " + String.join(";", codes));
                             String fieldKey = keyCache.computeIfAbsent(input.getNaaccrXmlId(), k -> String.valueOf(keyCache.size() + 1));
                             writer.write(keyCache.get(schemaId) + "," + fieldKey + "," + String.join(";", codes));
@@ -234,10 +232,10 @@ public class StagingDataGenerator {
 
     private static List<String> recombineRanges(List<String> list) {
         int length;
-        if (list.get(0).contains("-"))
-            length = StringUtils.split(list.get(0), '-')[0].length();
+        if (list.getFirst().contains("-"))
+            length = StringUtils.split(list.getFirst(), '-')[0].length();
         else
-            length = list.get(0).length();
+            length = list.getFirst().length();
 
         List<Integer> integers = new ArrayList<>();
         for (String s : list) {
@@ -245,11 +243,13 @@ public class StagingDataGenerator {
                 String[] parts = StringUtils.split(s, '-');
                 int low = Integer.parseInt(parts[0]);
                 int high = Integer.parseInt(parts[1]);
-                for (int i = low; i <= high; i++)
+                for (int i = low; i <= high; i++) {
                     integers.add(i);
+                }
             }
-            else
+            else {
                 integers.add(Integer.parseInt(s));
+            }
         }
 
         List<String> newRanges = new ArrayList<>();
@@ -257,19 +257,22 @@ public class StagingDataGenerator {
         int curIdx = 0;
         while (curIdx < integers.size()) {
             int nextIdx = curIdx + 1;
-            while (nextIdx < integers.size() && integers.get(nextIdx).equals(integers.get(nextIdx - 1) + 1))
+            while (nextIdx < integers.size() && integers.get(nextIdx).equals(integers.get(nextIdx - 1) + 1)) {
                 nextIdx++;
+            }
 
             Integer low = integers.get(curIdx);
             Integer high = integers.get(nextIdx - 1);
-            if (low.equals(high))
+            if (low.equals(high)) {
                 newRanges.add(StringUtils.leftPad(low.toString(), length, '0'));
+            }
             else if (low.equals(high - 1)) {
                 newRanges.add(StringUtils.leftPad(low.toString(), length, '0'));
                 newRanges.add(StringUtils.leftPad(high.toString(), length, '0'));
             }
-            else
+            else {
                 newRanges.add(StringUtils.leftPad(low.toString(), length, '0') + "-" + StringUtils.leftPad(high.toString(), length, '0'));
+            }
             curIdx = nextIdx;
         }