+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 **Version 2.3 (not released yet)**
 
+- Fix NAACCR sex rule to set the new sexAssignedAtBirth field for N26 and later.
 - Changed the minimum Java language for this library from Java 11 to Java 21.
 - Updated dependencies.